backend/requirements.txt
anthropic>=0.40.0
blinker==1.8.2
supabase==2.15.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
flask-cors==6.0.2
idna==3.7
importlib-metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.2.1
pytest==8.2.0
requests==2.31.0
tomli==2.0.1
typing-extensions==4.13.2
urllib3==2.2.1
werkzeug==3.0.3
zipp==3.18.1
